refactor init files
