+++ app/init/locales.php
@@ -1,0 +1,81 @@
+<?php
+
+use Utopia\Locale\Locale;
+
+
+/*
+ * Localization
+ */
+Locale::$exceptions = false;
+Locale::setLanguageFromJSON('af', __DIR__ . '../config/locale/translations/af.json');
+Locale::setLanguageFromJSON('ar', __DIR__ . '../config/locale/translations/ar.json');
+Locale::setLanguageFromJSON('as', __DIR__ . '../config/locale/translations/as.json');
+Locale::setLanguageFromJSON('az', __DIR__ . '../config/locale/translations/az.json');
+Locale::setLanguageFromJSON('be', __DIR__ . '../config/locale/translations/be.json');
+Locale::setLanguageFromJSON('bg', __DIR__ . '../config/locale/translations/bg.json');
+Locale::setLanguageFromJSON('bh', __DIR__ . '../config/locale/translations/bh.json');
+Locale::setLanguageFromJSON('bn', __DIR__ . '../config/locale/translations/bn.json');
+Locale::setLanguageFromJSON('bs', __DIR__ . '../config/locale/translations/bs.json');
+Locale::setLanguageFromJSON('ca', __DIR__ . '../config/locale/translations/ca.json');
+Locale::setLanguageFromJSON('cs', __DIR__ . '../config/locale/translations/cs.json');
+Locale::setLanguageFromJSON('da', __DIR__ . '../config/locale/translations/da.json');
+Locale::setLanguageFromJSON('de', __DIR__ . '../config/locale/translations/de.json');
+Locale::setLanguageFromJSON('el', __DIR__ . '../config/locale/translations/el.json');
+Locale::setLanguageFromJSON('en', __DIR__ . '../config/locale/translations/en.json');
+Locale::setLanguageFromJSON('eo', __DIR__ . '../config/locale/translations/eo.json');
+Locale::setLanguageFromJSON('es', __DIR__ . '../config/locale/translations/es.json');
+Locale::setLanguageFromJSON('fa', __DIR__ . '../config/locale/translations/fa.json');
+Locale::setLanguageFromJSON('fi', __DIR__ . '../config/locale/translations/fi.json');
+Locale::setLanguageFromJSON('fo', __DIR__ . '../config/locale/translations/fo.json');
+Locale::setLanguageFromJSON('fr', __DIR__ . '../config/locale/translations/fr.json');
+Locale::setLanguageFromJSON('ga', __DIR__ . '../config/locale/translations/ga.json');
+Locale::setLanguageFromJSON('gu', __DIR__ . '../config/locale/translations/gu.json');
+Locale::setLanguageFromJSON('he', __DIR__ . '../config/locale/translations/he.json');
+Locale::setLanguageFromJSON('hi', __DIR__ . '../config/locale/translations/hi.json');
+Locale::setLanguageFromJSON('hr', __DIR__ . '../config/locale/translations/hr.json');
+Locale::setLanguageFromJSON('hu', __DIR__ . '../config/locale/translations/hu.json');
+Locale::setLanguageFromJSON('hy', __DIR__ . '../config/locale/translations/hy.json');
+Locale::setLanguageFromJSON('id', __DIR__ . '../config/locale/translations/id.json');
+Locale::setLanguageFromJSON('is', __DIR__ . '../config/locale/translations/is.json');
+Locale::setLanguageFromJSON('it', __DIR__ . '../config/locale/translations/it.json');
+Locale::setLanguageFromJSON('ja', __DIR__ . '../config/locale/translations/ja.json');
+Locale::setLanguageFromJSON('jv', __DIR__ . '../config/locale/translations/jv.json');
+Locale::setLanguageFromJSON('kn', __DIR__ . '../config/locale/translations/kn.json');
+Locale::setLanguageFromJSON('km', __DIR__ . '../config/locale/translations/km.json');
+Locale::setLanguageFromJSON('ko', __DIR__ . '../config/locale/translations/ko.json');
+Locale::setLanguageFromJSON('la', __DIR__ . '../config/locale/translations/la.json');
+Locale::setLanguageFromJSON('lb', __DIR__ . '../config/locale/translations/lb.json');
+Locale::setLanguageFromJSON('lt', __DIR__ . '../config/locale/translations/lt.json');
+Locale::setLanguageFromJSON('lv', __DIR__ . '../config/locale/translations/lv.json');
+Locale::setLanguageFromJSON('ml', __DIR__ . '../config/locale/translations/ml.json');
+Locale::setLanguageFromJSON('mr', __DIR__ . '../config/locale/translations/mr.json');
+Locale::setLanguageFromJSON('ms', __DIR__ . '../config/locale/translations/ms.json');
+Locale::setLanguageFromJSON('nb', __DIR__ . '../config/locale/translations/nb.json');
+Locale::setLanguageFromJSON('ne', __DIR__ . '../config/locale/translations/ne.json');
+Locale::setLanguageFromJSON('nl', __DIR__ . '../config/locale/translations/nl.json');
+Locale::setLanguageFromJSON('nn', __DIR__ . '../config/locale/translations/nn.json');
+Locale::setLanguageFromJSON('or', __DIR__ . '../config/locale/translations/or.json');
+Locale::setLanguageFromJSON('pa', __DIR__ . '../config/locale/translations/pa.json');
+Locale::setLanguageFromJSON('pl', __DIR__ . '../config/locale/translations/pl.json');
+Locale::setLanguageFromJSON('pt-br', __DIR__ . '../config/locale/translations/pt-br.json');
+Locale::setLanguageFromJSON('pt-pt', __DIR__ . '../config/locale/translations/pt-pt.json');
+Locale::setLanguageFromJSON('ro', __DIR__ . '../config/locale/translations/ro.json');
+Locale::setLanguageFromJSON('ru', __DIR__ . '../config/locale/translations/ru.json');
+Locale::setLanguageFromJSON('sa', __DIR__ . '../config/locale/translations/sa.json');
+Locale::setLanguageFromJSON('sd', __DIR__ . '../config/locale/translations/sd.json');
+Locale::setLanguageFromJSON('si', __DIR__ . '../config/locale/translations/si.json');
+Locale::setLanguageFromJSON('sk', __DIR__ . '../config/locale/translations/sk.json');
+Locale::setLanguageFromJSON('sl', __DIR__ . '../config/locale/translations/sl.json');
+Locale::setLanguageFromJSON('sn', __DIR__ . '../config/locale/translations/sn.json');
+Locale::setLanguageFromJSON('sq', __DIR__ . '../config/locale/translations/sq.json');
+Locale::setLanguageFromJSON('sv', __DIR__ . '../config/locale/translations/sv.json');
+Locale::setLanguageFromJSON('ta', __DIR__ . '../config/locale/translations/ta.json');
+Locale::setLanguageFromJSON('te', __DIR__ . '../config/locale/translations/te.json');
+Locale::setLanguageFromJSON('th', __DIR__ . '../config/locale/translations/th.json');
+Locale::setLanguageFromJSON('tl', __DIR__ . '../config/locale/translations/tl.json');
+Locale::setLanguageFromJSON('tr', __DIR__ . '../config/locale/translations/tr.json');
+Locale::setLanguageFromJSON('uk', __DIR__ . '../config/locale/translations/uk.json');
+Locale::setLanguageFromJSON('ur', __DIR__ . '../config/locale/translations/ur.json');
+Locale::setLanguageFromJSON('vi', __DIR__ . '../config/locale/translations/vi.json');
+Locale::setLanguageFromJSON('zh-cn', __DIR__ . '../config/locale/translations/zh-cn.json');
+Locale::setLanguageFromJSON('zh-tw', __DIR__ . '../config/locale/translations/zh-tw.json');
